fix: remove whitespace normalization from a11y selectors
Previously, Puppeteer incorrectly normalized whitespaces in
ARIA selectors which does not allow distinguishing between
the following two HTML structures:

- <p>  text<p>
- <p><span>&nbsp;</span><span>&nbsp;</span>text<p>

In the first case, the spaces are not part of the element's
textual content, and the element should be found via `aria/text`.

In the second case, the spaces are part of the p element and the
query `aria/\u00A0\u00A0text` should find it. The whitespace normalization
in Puppeteer would previously prevent searching for the element with
whitespaces in the textual content.